+++ frameworks/Rust/salvo/Cargo.toml
@@ -15,14 +15,24 @@ path = "src/main_diesel.rs"
 name = "main-pg"
 path = "src/main_pg.rs"
 
+[[bin]]
+name = "cache-diesel"
+path = "src/cache_diesel.rs"
+
+[[bin]]
+name = "cache-pg"
+path = "src/cache_pg.rs"
+
 [dependencies]
 anyhow = "1"
 async-trait = "0.1"
 bytes = "1"
-diesel = { version = "1.4", features = ["postgres", "r2d2"] }
-futures = "0.3"
+diesel = { version = "2" }
+diesel-async = { version = "0.2", features = ["postgres", "deadpool"] }
+futures-util = "0.3"
+moka = "0.10"
 markup = "0.13"
-# mimalloc = { version = "0.1", default-features = false }
+mimalloc = { version = "0.1", default-features = false }
 once_cell = "1"
 rand = { version = "0.8", features = ["min_const_gen", "small_rng"] }
 random-fast-rng = "0.1"
