+++ frameworks/Rust/salvo/src/main_diesel.rs
@@ -1,20 +1,23 @@
 // #[global_allocator]
 // static ALLOC: snmalloc_rs::SnMalloc = snmalloc_rs::SnMalloc;
-// #[global_allocator]
-// static GLOBAL: mimalloc::MiMalloc = mimalloc::MiMalloc;
+#[global_allocator]
+static GLOBAL: mimalloc::MiMalloc = mimalloc::MiMalloc;
 
 #[macro_use]
 extern crate diesel;
 
 use std::cmp;
 use std::fmt::Write;
-use std::sync::Arc;
-use std::thread::available_parallelism;
 use std::net::{Ipv4Addr, SocketAddr};
+use std::sync::Arc;
 
 use anyhow::Error;
 use diesel::prelude::*;
-use diesel::r2d2::{ConnectionManager, Pool, PoolError, PooledConnection};
+
+use futures_util::stream::{FuturesUnordered, TryStreamExt};
+use diesel_async::pooled_connection::deadpool::Pool;
+use diesel_async::pooled_connection::AsyncDieselConnectionManager;
+use diesel_async::{AsyncPgConnection, RunQueryDsl};
 use once_cell::sync::OnceCell;
 use rand::rngs::SmallRng;
 use rand::{Rng, SeedableRng};
@@ -27,32 +30,33 @@ use models::*;
 use schema::*;
 
 const DB_URL: &str = "postgres://benchmarkdbuser:benchmarkdbpass@tfb-database/hello_world";
-type PgPool = Pool<ConnectionManager<PgConnection>>;
+type PgPool = Pool<AsyncPgConnection>;
 
 static DB_POOL: OnceCell<PgPool> = OnceCell::new();
-static CACHED_WORLDS: OnceCell<Vec<World>> = OnceCell::new();
 
-fn connect() -> Result<PooledConnection<ConnectionManager<PgConnection>>, PoolError> {
-    unsafe { DB_POOL.get_unchecked().get() }
+fn pool() -> &'static PgPool {
+    unsafe { DB_POOL.get_unchecked() }
 }
-fn build_pool(database_url: &str, size: u32) -> Result<PgPool, PoolError> {
-    let manager = ConnectionManager::<PgConnection>::new(database_url);
-    diesel::r2d2::Pool::builder()
+fn build_pool(database_url: &str, size: usize) -> PgPool {
+    let config = AsyncDieselConnectionManager::<AsyncPgConnection>::new(database_url);
+    Pool::builder(config)
         .max_size(size)
-        .min_idle(Some(size))
-        .test_on_check_out(false)
-        .idle_timeout(None)
-        .max_lifetime(None)
-        .build(manager)
+        .wait_timeout(None)
+        .build()
+        .unwrap_or_else(|_| panic!("Error connecting to {}", &DB_URL))
 }
 
 #[handler]
 async fn world_row(res: &mut Response) -> Result<(), Error> {
     let mut rng = SmallRng::from_entropy();
     let random_id = rng.gen_range(1..10_001);
-    let conn = connect()?;
-    let row = world::table.find(random_id).first::<World>(&conn)?;
-    res.headers_mut().insert(header::SERVER, HeaderValue::from_static("S"));
+    let mut conn = pool().get().await?;
+    let row = world::table
+        .find(random_id)
+        .first::<World>(&mut conn)
+        .await?;
+    res.headers_mut()
+        .insert(header::SERVER, HeaderValue::from_static("S"));
     res.render(Json(row));
     Ok(())
 }
@@ -61,33 +65,18 @@ async fn world_row(res: &mut Response) -> Result<(), Error> {
 async fn queries(req: &mut Request, res: &mut Response) -> Result<(), Error> {
     let count = req.query::<usize>("q").unwrap_or(1);
     let count = cmp::min(500, cmp::max(1, count));
-    let mut worlds = Vec::with_capacity(count);
-    let mut rng = SmallRng::from_entropy();
-    let conn = connect()?;
-    for _ in 0..count {
-        let id = rng.gen_range(1..10_001);
-        let w = world::table.find(id).get_result::<World>(&conn)?;
-        worlds.push(w);
-    }
-    res.headers_mut().insert(header::SERVER, HeaderValue::from_static("S"));
-    res.render(Json(worlds));
-    Ok(())
-}
-
-#[handler]
-async fn cached_queries(req: &mut Request, res: &mut Response) -> Result<(), Error> {
-    let count = req.query::<usize>("q").unwrap_or(1);
-    let count = cmp::min(500, cmp::max(1, count));
-    let mut worlds = Vec::with_capacity(count);
-    let mut rng = SmallRng::from_entropy();
-    for _ in 0..count {
-        let idx = rng.gen_range(0..10_000);
-        unsafe {
-            let w = CACHED_WORLDS.get_unchecked().get(idx).unwrap();
-            worlds.push(w);
-        }
-    }
-    res.headers_mut().insert(header::SERVER, HeaderValue::from_static("S"));
+    let worlds = {
+        let mut rng = SmallRng::from_entropy();
+        let mut conn = pool().get().await?;
+        (0..count)
+            .map(|_| {
+                let id = rng.gen_range(1..10_001);
+                world::table.find(id).first::<World>(&mut conn)
+            })
+            .collect::<FuturesUnordered<_>>()
+    }.try_collect::<Vec<_>>().await?;
+    res.headers_mut()
+        .insert(header::SERVER, HeaderValue::from_static("S"));
     res.render(Json(worlds));
     Ok(())
 }
@@ -96,35 +85,49 @@ async fn cached_queries(req: &mut Request, res: &mut Response) -> Result<(), Err
 async fn updates(req: &mut Request, res: &mut Response) -> Result<(), Error> {
     let count = req.query::<usize>("q").unwrap_or(1);
     let count = cmp::min(500, cmp::max(1, count));
-    let conn = connect()?;
-    let mut worlds = Vec::with_capacity(count);
-    let mut rng = SmallRng::from_entropy();
-    for _ in 0..count {
-        let w_id: i32 = rng.gen_range(1..10_001);
-        let mut w = world::table.find(w_id).first::<World>(&conn)?;
-        w.randomnumber = rng.gen_range(1..10_001);
-        worlds.push(w);
+    let mut conn = pool().get().await?;
+    let mut worlds = {
+        let mut rng = SmallRng::from_entropy();
+        (0..count)
+            .map(|_| {
+                let w_id: i32 = rng.gen_range(1..10_001);
+                let new_id = rng.gen_range(1..10_001);
+                let fut = world::table.find(w_id).first::<World>(&mut conn);
+                async move {
+                    let mut w = fut.await?;
+                    w.randomnumber = new_id;
+                    Result::<_, Error>::Ok(w)
+                }
+            })
+            .collect::<FuturesUnordered<_>>()
     }
+    .try_collect::<Vec<_>>()
+    .await?;
     worlds.sort_by_key(|w| w.id);
-    conn.transaction::<(), Error, _>(|| {
-        for w in &worlds {
-            diesel::update(world::table)
-                .filter(world::id.eq(w.id))
-                .set(world::randomnumber.eq(w.randomnumber))
-                .execute(&conn)?;
-        }
-        Ok(())
-    })?;
+    conn.build_transaction()
+        .run(|conn| {
+            Box::pin(async {
+                for w in &worlds {
+                    diesel::update(world::table.find(w.id))
+                        .set(world::randomnumber.eq(w.randomnumber))
+                        .execute(conn)
+                        .await?;
+                }
+                Ok::<(), Error>(())
+            })
+        })
+        .await?;
 
-    res.headers_mut().insert(header::SERVER, HeaderValue::from_static("S"));
+    res.headers_mut()
+        .insert(header::SERVER, HeaderValue::from_static("S"));
     res.render(Json(worlds));
     Ok(())
 }
 
 #[handler]
 async fn fortunes(res: &mut Response) -> Result<(), Error> {
-    let conn = connect()?;
-    let mut items = fortune::table.get_results::<Fortune>(&conn)?;
+    let mut conn = pool().get().await?;
+    let mut items = fortune::table.get_results::<Fortune>(&mut conn).await?;
     items.push(Fortune {
         id: 0,
         message: "Additional fortune added at request time.".to_string(),
@@ -134,7 +137,8 @@ async fn fortunes(res: &mut Response) -> Result<(), Error> {
     let mut body = String::new();
     write!(&mut body, "{}", FortunesTemplate { items }).unwrap();
 
-    res.headers_mut().insert(header::SERVER, HeaderValue::from_static("S"));
+    res.headers_mut()
+        .insert(header::SERVER, HeaderValue::from_static("S"));
     res.render(Text::Html(body));
     Ok(())
 }
@@ -161,13 +165,6 @@ markup::define! {
     }
 }
 
-fn populate_cache() -> Result<(), Error> {
-    let conn = connect()?;
-    let worlds = world::table.limit(10_000).get_results::<World>(&conn)?;
-    CACHED_WORLDS.set(worlds).unwrap();
-    Ok(())
-}
-
 #[tokio::main]
 async fn main() {
     let router = Arc::new(
@@ -175,14 +172,9 @@ async fn main() {
             .push(Router::with_path("db").get(world_row))
             .push(Router::with_path("fortunes").get(fortunes))
             .push(Router::with_path("queries").get(queries))
-            .push(Router::with_path("cached_queries").get(cached_queries))
             .push(Router::with_path("updates").get(updates)),
     );
-    let size = available_parallelism().map(|n| n.get()).unwrap_or(16);
-    DB_POOL
-        .set(build_pool(DB_URL, size as u32).unwrap_or_else(|_| panic!("Error connecting to {}", &DB_URL)))
-        .ok();
-    populate_cache().expect("error cache worlds");
+    DB_POOL.set(build_pool(DB_URL, 5)).ok();
 
     let addr = SocketAddr::from((Ipv4Addr::UNSPECIFIED, 8080));
     let acceptor = TcpListener::new(addr).bind().await;
